requirements.txt
Django
djangorestframework
python-jose
